requirements.txt
 Flask==2.3.3
py2neo==2021.2.4
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
openpyxl==3.1.2
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0